@@ -1,5 +1,5 @@
 class CmsController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   
   def index
     bring_in_models
