@@ -1,5 +1,5 @@
 class ContactsController < ApplicationController
-  before_filter :authenticate_user!, :except => ["index"]
+  before_action :authenticate_user!, :except => ["index"]
   
   def index
     bring_in_models
