@@ -1,8 +1,6 @@
 V1::Application.routes.draw do
-  ActiveAdmin.routes(self)
 
   devise_for :users, path: "auth", path_names: { sign_in: 'login', sign_out: 'logout', sign_up: 'create_new_splitter' }
-
   root :to => 'announcements#index'
   
   resources :cms
@@ -24,6 +22,6 @@ V1::Application.routes.draw do
   resources :abouts, :path => "about"
   resources :contacts, :path => "contact"
   
-  match '*path' => redirect('/')
+  get '*path' => redirect('/')
   
 end
